api/pyproject.toml
[project]
name = "translator-api"
version = "0.1.0"
description = "FastAPI backend for the LSTM EN→ES translator"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic-settings>=2.0",
    "translator",
    "huggingface-hub>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/api"]

[tool.uv.sources]
translator = { path = "../lstm-translator", editable = true }
torch = [{ index = "pytorch-cu130" }]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
